fix warning message

+++ ObjectiveCExample/ObjectiveCExampleTests/SSZipArchiveTests.m
@@ -23,7 +23,7 @@
 @implementation CancelDelegate
 - (void)zipArchiveDidUnzipFileAtIndex:(NSInteger)fileIndex totalFiles:(NSInteger)totalFiles archivePath:(NSString *)archivePath fileInfo:(unz_file_info)fileInfo
 {
-    _numFilesUnzipped = fileIndex + 1;
+    _numFilesUnzipped = (int)(fileIndex + 1);
 }
 - (BOOL)zipArchiveShouldUnzipFileAtIndex:(NSInteger)fileIndex totalFiles:(NSInteger)totalFiles archivePath:(NSString *)archivePath fileInfo:(unz_file_info)fileInfo
 {
